requirements.txt
Flask==3.1.1
mysql-connector-python==9.3.0
bcrypt==4.1.3
Flask-JWT-Extended